+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
+## v1.3.6
+- change: Add FAQ link to toolbar.
+- feat: add new option to hide unselected sources in rune popup.
+
 ## v1.3.5
 - change: check lcu status task
-- style(popup): rune list 
+- style(popup): rune list
 
 ## v1.3.4
 - fix: correct source list url